Refactor redcap_det_uw_reopening move to redcap.py
